Fix UD-327 lodash instead of underscore
Lodash is wiser than underscore for iterating over AngularJS objects.
For instance, in deferred/promises that resolve to arrays, only array
is iterated, not deferred/promises related properties. Moreover lodash
is now a superset of underscore providing thus the same features but
in a better way (see https://stackoverflow.com/a/13898916/256561 for
explanation of the author himself).